+++ src/components/Heroes/Heroes.css
@@ -1,0 +1,3 @@
+.Heading {
+  composes: Heading from "../Distributions/Distributions.css";
+}
