+++ src/lang/en-US.json
@@ -244,7 +244,9 @@
   "heading_region": "Region",
   "heading_patch": "Patch",
   
-  "hero_this_month": "Last 30 days",
+  "hero_pro_heading": "Picks & Bans in Professional Matches",
+  "hero_ranked_heading": "Picks & Bans in Ranked Matches",
+  "hero_this_month": "matches this month",
   "hero_pick_ban_rate": "Pro P+B%",
   "hero_pick_rate": "Pro Pick%",
   "hero_ban_rate": "Pro Ban%",
